requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
paho-mqtt==1.6.1
python-engineio==4.7.1
python-socketio==5.9.0
google-generativeai==0.3.2
python-dotenv==1.0.0 